// src/gateways/jubmojis/jubmoji.types.ts
export interface MembershipProof {
  R: EdwardsPoint;
  msgHash: bigint;
  zkp: ZKP;
}
// ...
export interface MerkleRootArgs {
  pubKeys: WeierstrassPoint[];
  hashFn?: any;
}
// ...
export interface VerifyArgs {
  proof: MembershipProof;
  merkleRoot?: bigint;
  merkleRootArgs?: MerkleRootArgs;
  sigNullifierRandomness: bigint;
  usedSigNullifiers?: bigint[];
  pathToCircuits?: string;
  enableTiming?: boolean;
}

// Arguments needed to batch verify membership proofs
export interface BatchVerifyArgs {
  proofs: MembershipProof[];
  merkleRoot?: bigint;
  merkleRootArgs?: MerkleRootArgs;
  sigNullifierRandomness: bigint;
  usedSigNullifiers?: bigint[];
  pathToCircuits?: string;
  enableTiming?: boolean;
}

export type VerificationResult = {
  verified: boolean;
  consumedSigNullifiers?: bigint[];
};

// Zero knowledge proof generated by snarkjs
export type ZKP = { proof: any; publicSignals: string[] };
// ...
export type ZKPPublicSignals = {
  merkleRoot: bigint;
  T: EdwardsPoint;
  U: EdwardsPoint;
  sigNullifier: bigint;
  sigNullifierRandomness: bigint;
  pubKeyNullifier: bigint;
  pubKeyNullifierRandomnessHash: bigint;
};